config/requirements/requirements.txt
# Основные зависимости для пайплайна машинного обучения
# apache-airflow==2.8.1 # Уже установлен в базовом образе
apache-airflow==2.10.5 # Поддерживает 3.12
pandas==2.1.4
numpy==1.24.4 # Совместимо с pandas 2.1.4
scikit-learn==1.3.2
scipy==1.11.4
joblib==1.3.2
python-dotenv==1.0.0
pydantic==1.10.13 # Более старая версия для совместимости
pyyaml==6.0.1

# Визуализация и анализ данных
matplotlib==3.7.4 # Совместимо с numpy 1.24.4
seaborn==0.12.2

# Облачные хранилища
google-cloud-storage==2.10.0
boto3==1.34.0

# Базы данных
sqlalchemy>=1.4.0,<2.0.0 # Совместимо с Apache Airflow 2.10.5
psycopg2-binary==2.9.9
pymysql==1.1.0
# sqlite3 встроена в Python

# Логирование и мониторинг
structlog==23.2.0

# Тестирование
pytest==7.4.3
pytest-cov==4.1.0

# Линтинг и форматирование кода
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Дополнительные инструменты для анализа данных
plotly==5.17.0
imbalanced-learn==0.11.0
